python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "databricks-workspace-client"
version = "0.1.0"
description = "Reusable client library wrapping the Databricks SDK for workspace operations"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "databricks-sdk>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "ruff>=0.4.0",
    "mypy>=1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/databricks_workspace_client"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["databricks_workspace_client"]
mypy_path = "src"
